# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LaunchMailer API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/setup:
    post:
      operationId: setupAdmin
      tags: [auth]
      summary: Setup initial admin account (only works if no admin exists)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupAdminInput"
      responses:
        "201":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Admin already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"

    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign details with logs
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign with logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── UPLOAD ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # ─── AUTH ─────────────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
          minLength: 1
    SetupAdminInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string

        password:
          type: string
          minLength: 8
    Admin:
      type: object
      required: [id, email, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [token, admin]
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/Admin"

    # ─── DASHBOARD ────────────────────────────────────────────────────────────
...
    Campaign:
      type: object
      required: [id, name, method, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        subject:
          type: ["string", "null"]
        emailContent:
          type: ["string", "null"]
        whatsappMessage:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        method:
          type: string
          enum: [email, whatsapp, both]
        status:
          type: string
          enum: [draft, sending, sent, failed]
        recipientCount:
          type: integer
        emailSent:
          type: integer
        whatsappSent:
          type: integer
        successCount:
          type: integer
        failedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        sentAt:
          type: ["string", "null"]
          format: date-time
    CampaignInput:
      type: object
      required: [name, method]
      properties:
        name:
          type: string
          minLength: 1
        subject:
          type: string
        emailContent:
          type: string
        whatsappMessage:
          type: string
        imageUrl:
          type: string
        method:
          type: string
          enum: [email, whatsapp, both]
        recipientFilter:
          $ref: "#/components/schemas/RecipientFilter"
    RecipientFilter:
      type: object
      properties:
        hasEmail:
          type: boolean
        hasPhone:
          type: boolean
        city:
          type: string
        tags:
          type: array
          items:
            type: string
    CampaignLog:
      type: object
      required: [id, channel, status, createdAt]
      properties:
        id:
          type: string
        customerId:
          type: string
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        channel:
          type: string
          enum: [email, whatsapp]
        status:
          type: string
          enum: [sent, delivered, failed, opened]
        messageId:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CampaignDetail:
      allOf:
        - $ref: "#/components/schemas/Campaign"
        - type: object
          required: [logs]
          properties:
            logs:
              type: array
              items:
                $ref: "#/components/schemas/CampaignLog"

    # ─── UPLOAD ───────────────────────────────────────────────────────────────
...