# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * LaunchMailer API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: string;
  email: string;
  createdAt: Date;
}